carpark live data cache and load task to insert all live data is done. lots of cleanup to rename classes

+++ src/main/java/com/wego/assignment/common/view/LatLong.java
@@ -2,22 +2,22 @@ package com.wego.assignment.common.view;
 
 public class LatLong {
 
-    private  double latitude;
-    private  double longitude;
+    private  Double latitude;
+    private  Double longitude;
 
-    public double getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public double getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
